rss feed plus
Added an action for creating an RSS 1.0 (RDF) feed. To make this work,
added a couple of helper functions to util, and changed the HTML stuff
to use the helper functions.

Also, changed a bit of formatting in the profile. Yeah, not a clean
changeset.

darcs-hash:20080521143351-84dde-d191761c399d7f014b990585c21e916e887b3401.gz