tags and XRDS
Starting the process of remote subscriptions.

First, add the capacity to mint tags. May in the future want to allow
some kind of UUID or URL for unique URIs. New config options for tag
authority and date, and a function for minting tags.

Store URIs in the DB. If the admin changes his/her tag config, this
shouldn't change the URIs in the DB. Unless they really want to make
that happen.

Generate an XRDS document with the appropriate URLs for microblogging.
Also, redefined some stuff in openmicroblogging.txt. And moved the file.

darcs-hash:20080522183421-84dde-fe46f32e4050f2c4b6c850334103007ed7feabf7.gz