Update language q-factors to represent current translated values
I've updated the q-factors in lib/language.php to represent the
percentage of actual translated values in the files, per Pootle. It
may make sense in the future to weight this to favor "important"
translations, e.g. stuff that shows up more frequently rather than in
rare error messages.

I've also added a few languages that weren't in the file.