Move page-generation code from lib/util.php to Action class
We moved the page generation code from lib/util.php to the action class.

We also broke up the page generation code into a big recipe, where
each element of the page has a corresponding method in the Action
class. Action classes can override parts of the recipe to make their
custom output, and leave the rest up to the parent class.

NOTE: this change BREAKS pretty much everything. lib/action.php probably
does not compile, and everything that uses it needs to be fixed to use
the new framework. Stay tuned for more fun!